refactor: remove business signup
