+++ frontend/src/router.ts
@@ -11,21 +11,9 @@ const routes = [
 		meta: { isLogin: true },
 	},
 	{
-		path: '/signup/personal',
-		name: 'PersonalSignUp',
-		component: () => import('@/pages/SignupView.vue'),
-		meta: { isLogin: true },
-	},
-	{
-		path: '/signup/business',
-		name: 'BusinessSignUp',
-		component: () => import('@/pages/BusinessSignupView.vue'),
-		meta: { isLogin: true },
-	},
-	{
-		path: '/signup/business/:requestKey',
-		name: 'BusinessSetup',
-		component: () => import('@/pages/BusinessSignupView.vue'),
+		path: '/signup/:requestKey',
+		name: 'InviteSetup',
+		component: () => import('@/pages/InviteSetupView.vue'),
 		props: true,
 		meta: { isLogin: true },
 	},
