--- mail/templates/emails/business_signup.html
@@ -1,24 +1,0 @@
-{% import "templates/emails/macros.html" as utils %}
-{% extends "templates/emails/base.html" %}
-{% set otp = otp %}
-
-{% block content %}
-
-<td class="text-base leading-6 text-gray-900">
-    <div class="p-8 bg-white from-markdown">
-        <p><strong>You're almost done!</strong></p>
-        <p>Just one quick step left to get you started with Frappe Mail!</p>
-        {% if otp %}
-        <p class="font-medium">Verification Code</p>
-        <div class="text-xl font-bold">{{ otp }}</div>
-        <p>Or click on the button to verify your account</p>
-        {% else %}
-        <p>Click on the button to verify your account</p>
-        {% endif %}
-        {{ utils.button('Verify Account', link) }}
-        {{ utils.separator() }}
-        <p class="my-0">Team Frappe</p>
-    </div>
-</td>
-
-{% endblock %}
